#1047 Extracted static methods from IValueMatcher into ValueMatcherUtils. Unified implementation of 'createCommonQueryBuilderForFindMatches' in AbstractSearchEntityByKeyWithContext and AbstractSearchEntityByKeyWithCentreContext in preparation to replace it with a single implementation. Removed the default page size in favor of the default implementation for 'getPageSize' in IValueMatcher.
