+++ platform-pojo-bl/src/main/java/ua/com/fielden/platform/basic/IValueMatcher.java
@@ -25,5 +25,7 @@ public interface IValueMatcher<T> {
      *
      * @return
      */
-    Integer getPageSize();
+    default Integer getPageSize() {
+        return 10;
+    }
 }
