+++ platform-pojo-bl/src/main/java/ua/com/fielden/platform/basic/autocompleter/AbstractSearchEntityByKeyWithCentreContext.java
@@ -3,16 +3,16 @@
  */
 package ua.com.fielden.platform.basic.autocompleter;
 
+import static java.util.Collections.emptyMap;
+import static ua.com.fielden.platform.basic.ValueMatcherUtils.createRelaxedSearchByKeyCriteriaModel;
 import static ua.com.fielden.platform.entity.AbstractEntity.KEY;
 import static ua.com.fielden.platform.entity.query.fluent.EntityQueryUtils.cond;
 import static ua.com.fielden.platform.entity.query.fluent.EntityQueryUtils.fetchKeyAndDescOnly;
 import static ua.com.fielden.platform.entity.query.fluent.EntityQueryUtils.from;
 import static ua.com.fielden.platform.entity.query.fluent.EntityQueryUtils.orderBy;
 import static ua.com.fielden.platform.entity.query.fluent.EntityQueryUtils.select;
-import static ua.com.fielden.platform.basic.IValueMatcherWithFetch.createRelaxedSearchByKeyCriteriaModel;
 import static ua.com.fielden.platform.utils.EntityUtils.hasDescProperty;
 
-import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.Optional;
@@ -35,18 +35,14 @@ import ua.com.fielden.platform.web.centre.CentreContext;
  *
  * @author TG Team
  */
-public abstract class AbstractSearchEntityByKeyWithCentreContext<T extends AbstractEntity<?>>
-                      implements IValueMatcherWithCentreContext<T>, IValueMatcherWithFetch<T> {
+public abstract class AbstractSearchEntityByKeyWithCentreContext<T extends AbstractEntity<?>> implements IValueMatcherWithCentreContext<T>, IValueMatcherWithFetch<T> {
 
-    private static final int DEFAULT_PAGE_SIZE = 10;
     private static final Supplier<? extends EntityCentreExecutionException> CO_MISSING_EXCEPTION_SUPPLIER = () -> new EntityCentreExecutionException("A companion is massing to perform this operation.");
 
     private final Optional<IEntityDao<T>> maybeCompanion;
     private final fetch<T> defaultFetchModel;
     private fetch<T> fetchModel;
     private CentreContext<T, ?> context;
-
-
 
     public AbstractSearchEntityByKeyWithCentreContext(final IEntityDao<T> companion) {
         this.maybeCompanion = Optional.ofNullable(companion);
@@ -61,19 +57,18 @@ public abstract class AbstractSearchEntityByKeyWithCentreContext<T extends Abstr
      * @return
      */
     protected ConditionModel makeSearchCriteriaModel(final CentreContext<T, ?> context, final String searchString) {
-    	if ("%".equals(searchString)) {
-    		return cond().val(1).eq().val(1).model();
-    	}
-      	
-    	final ConditionModel keyCriteria = createRelaxedSearchByKeyCriteriaModel(searchString);
+        if ("%".equals(searchString)) {
+            return cond().val(1).eq().val(1).model();
+        }
+
+        final ConditionModel keyCriteria = createRelaxedSearchByKeyCriteriaModel(searchString);
         final Class<T> entityType = maybeCompanion.orElseThrow(CO_MISSING_EXCEPTION_SUPPLIER).getEntityType();
-      
-    	return hasDescProperty(entityType) ? cond().condition(keyCriteria).or().prop(AbstractEntity.DESC).iLike().val("%" + searchString).model() : keyCriteria;
+
+        return hasDescProperty(entityType) ? cond().condition(keyCriteria).or().prop(AbstractEntity.DESC).iLike().val("%" + searchString).model() : keyCriteria;
     }
 
     /**
-     * This method may be overridden to provide the param values for the
-     * resulting query based on the provided context.
+     * This method may be overridden to provide the param values for the resulting query based on the provided context.
      *
      * @param context
      * @param params
@@ -84,8 +79,7 @@ public abstract class AbstractSearchEntityByKeyWithCentreContext<T extends Abstr
     }
 
     /**
-     * This method may be overridden to provide an alternative ordering if the
-     * default ordering by the key is not suitable.
+     * This method may be overridden to provide an alternative ordering if the default ordering by the key is not suitable.
      *
      * @return alternative ordering model
      */
@@ -95,20 +89,16 @@ public abstract class AbstractSearchEntityByKeyWithCentreContext<T extends Abstr
 
     private Builder<T, EntityResultQueryModel<T>> createCommonQueryBuilderForFindMatches(final String searchString, final Class<T> entityType) {
         final ConditionModel searchCriteria = makeSearchCriteriaModel(getContext(), searchString);
-        final EntityResultQueryModel<T> queryModel = searchCriteria != null ? select(entityType).where().condition(searchCriteria).model() : select(entityType).model();
-        queryModel.setFilterable(true);
+        final EntityResultQueryModel<T> queryModel = (searchCriteria != null ? select(entityType).where().condition(searchCriteria).model() : select(entityType).model()).setFilterable(true);
         final OrderingModel ordering = composeOrderingModelForQuery(searchString, entityType);
-        final Map<String, Object> params = new HashMap<>();
-        fillParamsBasedOnContext(getContext(), params);
-        return from(queryModel).with(ordering).with(params).lightweight();
+        
+        fillParamsBasedOnContext(getContext(), emptyMap());
+        return from(queryModel).with(ordering).with(emptyMap()).lightweight();
     }
-    
+
     private OrderingModel composeOrderingModelForQuery(final String searchString, final Class<T> entityType) {
-    	return "%".equals(searchString) ? makeOrderingModel(searchString) : 
-        	orderBy().
-        	expr(makeSearchResultOrderingPriority(entityType, searchString)).asc().
-        	order(makeOrderingModel(searchString)).
-        	model();
+        return "%".equals(searchString) ? makeOrderingModel(searchString)
+                : orderBy().expr(makeSearchResultOrderingPriority(entityType, searchString)).asc().order(makeOrderingModel(searchString)).model();
     }
 
     @Override
@@ -144,8 +134,4 @@ public abstract class AbstractSearchEntityByKeyWithCentreContext<T extends Abstr
         return this;
     }
 
-    @Override
-    public Integer getPageSize() {
-        return DEFAULT_PAGE_SIZE;
-    }
 }
