+++ platform-pojo-bl/src/main/java/ua/com/fielden/platform/basic/autocompleter/AbstractSearchEntityByKeyWithContext.java
@@ -3,6 +3,8 @@
  */
 package ua.com.fielden.platform.basic.autocompleter;
 
+import static java.util.Collections.emptyMap;
+import static ua.com.fielden.platform.basic.ValueMatcherUtils.createRelaxedSearchByKeyCriteriaModel;
 import static ua.com.fielden.platform.entity.AbstractEntity.DESC;
 import static ua.com.fielden.platform.entity.AbstractEntity.KEY;
 import static ua.com.fielden.platform.entity.query.fluent.EntityQueryUtils.cond;
@@ -11,7 +13,6 @@ import static ua.com.fielden.platform.entity.query.fluent.EntityQueryUtils.from;
 import static ua.com.fielden.platform.entity.query.fluent.EntityQueryUtils.orderBy;
 import static ua.com.fielden.platform.entity.query.fluent.EntityQueryUtils.select;
 import static ua.com.fielden.platform.utils.EntityUtils.hasDescProperty;
-import static ua.com.fielden.platform.basic.IValueMatcherWithFetch.createRelaxedSearchByKeyCriteriaModel;
 
 import java.util.HashMap;
 import java.util.List;
@@ -32,16 +33,12 @@ import ua.com.fielden.platform.entity.query.model.OrderingModel;
  *
  * @author TG Team
  */
-public abstract class AbstractSearchEntityByKeyWithContext<CONTEXT extends AbstractEntity<?>, T extends AbstractEntity<?>>
-                      implements IValueMatcherWithContext<CONTEXT, T>, IValueMatcherWithFetch<T> {
+public abstract class AbstractSearchEntityByKeyWithContext<CONTEXT extends AbstractEntity<?>, T extends AbstractEntity<?>> implements IValueMatcherWithContext<CONTEXT, T>, IValueMatcherWithFetch<T> {
 
     protected final IEntityDao<T> companion;
     private final fetch<T> defaultFetchModel;
     private fetch<T> fetchModel;
     private CONTEXT context;
-
-    private final int pageSize = 10;
-
 
     public AbstractSearchEntityByKeyWithContext(final IEntityDao<T> companion) {
         this.companion = companion;
@@ -56,18 +53,17 @@ public abstract class AbstractSearchEntityByKeyWithContext<CONTEXT extends Abstr
      * @return
      */
     protected ConditionModel makeSearchCriteriaModel(final CONTEXT context, final String searchString) {
-    	if ("%".equals(searchString)) {
-    		return cond().val(1).eq().val(1).model();
-    	}
-      	
-    	final ConditionModel keyCriteria = createRelaxedSearchByKeyCriteriaModel(searchString);
-    	
-    	return hasDescProperty(companion.getEntityType()) ? cond().condition(keyCriteria).or().prop(DESC).iLike().val("%" + searchString).model() : keyCriteria;
+        if ("%".equals(searchString)) {
+            return cond().val(1).eq().val(1).model();
+        }
+
+        final ConditionModel keyCriteria = createRelaxedSearchByKeyCriteriaModel(searchString);
+
+        return hasDescProperty(companion.getEntityType()) ? cond().condition(keyCriteria).or().prop(DESC).iLike().val("%" + searchString).model() : keyCriteria;
     }
 
     /**
-     * This method may be overridden to provide the param values for the
-     * resulting query based on the provided context.
+     * This method may be overridden to provide the param values for the resulting query based on the provided context.
      *
      * @param context
      * @param params
@@ -78,8 +74,7 @@ public abstract class AbstractSearchEntityByKeyWithContext<CONTEXT extends Abstr
     }
 
     /**
-     * This method may be overridden to provide an alternative ordering if the
-     * default ordering by the key is not suitable.
+     * This method may be overridden to provide an alternative ordering if the default ordering by the key is not suitable.
      *
      * @return alternative ordering model
      */
@@ -87,32 +82,28 @@ public abstract class AbstractSearchEntityByKeyWithContext<CONTEXT extends Abstr
         return orderBy().prop(KEY).asc().model();
     }
 
-    private Builder<T, EntityResultQueryModel<T>> createCommonQueryBuilderForFindMatches(final String searchString) {
+    private Builder<T, EntityResultQueryModel<T>> createCommonQueryBuilderForFindMatches(final String searchString, final Class<T> entityType) {
         final ConditionModel searchCriteria = makeSearchCriteriaModel(getContext(), searchString);
-        final EntityResultQueryModel<T> queryModel = searchCriteria != null ? select(companion.getEntityType()).where().condition(searchCriteria).model() : select(companion.getEntityType()).model();
-        queryModel.setFilterable(true);
+        final EntityResultQueryModel<T> queryModel = (searchCriteria != null ? select(entityType).where().condition(searchCriteria).model() : select(entityType).model()).setFilterable(true);
         final OrderingModel ordering = composeOrderingModelForQuery(searchString);
-        final Map<String, Object> params = new HashMap<>();
-        fillParamsBasedOnContext(getContext(), params);
-        return from(queryModel).with(ordering).with(params).lightweight();
+
+        fillParamsBasedOnContext(getContext(), emptyMap());
+        return from(queryModel).with(ordering).with(emptyMap()).lightweight();
     }
-    
+
     private OrderingModel composeOrderingModelForQuery(final String searchString) {
-    	return "%".equals(searchString) ? makeOrderingModel(searchString) : 
-        	orderBy().
-        	expr(makeSearchResultOrderingPriority(companion.getEntityType(), searchString)).asc().
-        	order(makeOrderingModel(searchString)).
-        	model();
+        return "%".equals(searchString) ? makeOrderingModel(searchString)
+                : orderBy().expr(makeSearchResultOrderingPriority(companion.getEntityType(), searchString)).asc().order(makeOrderingModel(searchString)).model();
     }
 
     @Override
     public List<T> findMatches(final String searchString) {
-        return companion.getFirstEntities(createCommonQueryBuilderForFindMatches(searchString).with(defaultFetchModel).model(), getPageSize());
+        return companion.getFirstEntities(createCommonQueryBuilderForFindMatches(searchString, companion.getEntityType()).with(defaultFetchModel).model(), getPageSize());
     }
 
     @Override
     public List<T> findMatchesWithModel(final String searchString) {
-        return companion.getFirstEntities(createCommonQueryBuilderForFindMatches(searchString).with(getFetch()).model(), getPageSize());
+        return companion.getFirstEntities(createCommonQueryBuilderForFindMatches(searchString, companion.getEntityType()).with(getFetch()).model(), getPageSize());
     }
 
     @Override
@@ -135,8 +126,4 @@ public abstract class AbstractSearchEntityByKeyWithContext<CONTEXT extends Abstr
         this.context = context;
     }
 
-    @Override
-    public Integer getPageSize() {
-        return pageSize;
-    }
 }
